requirements.txt
# Core dependencies
streamlit>=1.31.0
anthropic>=0.8.0
Pillow>=10.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0
pygments>=2.17.2
firebase-admin>=6.4.0
pytz>=2024.1
tzdata>=2024.1

# Git integration
GitPython>=3.1.40

# Formatting and analysis
black>=23.3.0

# Visualization
matplotlib>=3.8.0
plotly>=5.18.0

# Development tools
pytest>=7.4.0
pytest-mock>=3.10.0
pytest-cov>=4.1.0
pylint>=3.0.0
isort>=5.12.0

# Utility
python-slugify>=8.0.1
pydantic>=2.6.1
requests>=2.31.0
typing-extensions>=4.7.1
dataclasses-json>=0.6.1